+++ packages/cli/CHANGELOG.md
@@ -23,6 +23,8 @@
   - Updated `moon bin` and `moon docker` commands to support Bun.
   - Updated task `platform` to support "bun".
 - Improved TypeScript support.
+  - Added a `typescript.includeSharedTypes` setting, for syncing a shared types path to all project
+    `include`.
   - Updated the `typescript.rootConfigFileName` setting to support directories.
 
 ## 1.16.5